finetune/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
tqdm==4.66.4
datasets==2.21.0
